+++ modules/nameplates.lua
@@ -1,15 +1,15 @@
 function GigaSettingsInterface:CastTimerNameplate(nameplate)
-  if not nameplate:IsForbidden() and not nameplate.UnitFrame.castBar.timer then
-    nameplate.UnitFrame.castBar.timer = nameplate.UnitFrame.castBar:CreateFontString(nil, "OVERLAY", "GameFontHighlightSmall")
-    local fontPath = nameplate.UnitFrame.castBar.timer:GetFont()
-    nameplate.UnitFrame.castBar.timer:SetFont(fontPath, 8, "OUTLINE")
-    nameplate.UnitFrame.castBar.timer:SetPoint("LEFT", nameplate.UnitFrame.castBar, "RIGHT", -35, 0)
-    nameplate.UnitFrame.castBar.timer:SetText("")
-    nameplate.UnitFrame.castBar:HookScript("OnValueChanged", function(self)
+  if not nameplate:IsForbidden() and not nameplate.UnitFrame.CastBarsContainer.castBar.timer then
+    nameplate.UnitFrame.CastBarsContainer.castBar.timer = nameplate.UnitFrame.CastBarsContainer.castBar:CreateFontString(nil, "OVERLAY", "GameFontHighlightSmall")
+    local fontPath = nameplate.UnitFrame.CastBarsContainer.castBar.timer:GetFont()
+    nameplate.UnitFrame.CastBarsContainer.castBar.timer:SetFont(fontPath, 13, "OUTLINE, SLUG")
+    nameplate.UnitFrame.CastBarsContainer.castBar.timer:SetPoint("RIGHT", nameplate.UnitFrame.CastBarsContainer.castBar, "RIGHT", 0, 0)
+    nameplate.UnitFrame.CastBarsContainer.castBar.timer:SetText("")
+    nameplate.UnitFrame.CastBarsContainer.castBar:HookScript("OnValueChanged", function(self)
       if self.casting then
-        self.timer:SetText(format("%2.1f/%1.1f", nameplate.UnitFrame.castBar.value, nameplate.UnitFrame.castBar.maxValue))
+        self.timer:SetText(format("%2.1f/%1.1f", self:GetValue(), select(2, self:GetMinMaxValues())))
       elseif self.channeling then
-        self.timer:SetText(format("%.1f", nameplate.UnitFrame.castBar.value))
+        self.timer:SetText(format("%.1f", self:GetValue()))
       else
         self.timer:SetText("")
       end